+++ CHANGES.md
@@ -33,8 +33,9 @@
     - @sile
 - [FIX] クライアント名が消えてたのを修正
     - @melpon
-- [FIX] パッケージを展開した際の名前が `zakuro` になってしまっていたのを `zakuro-{version}_{platform}` にする
+- [FIX] パッケージを展開した際の名前が `zakuro` になってしまっていたのを `zakuro-{version}` にする
     - @melpon
+    - @miosakuma
 
 ## 2022.7.1 (2022-10-31)