+++ CHANGES.md
@@ -15,7 +15,7 @@
     - @melpon
 - [UPDATE] WebRTC を m111.5563.4.4 に上げる
     - @melpon
-- [UPDATE] Sora C++ SDK を 2023.3.0 に上げる
+- [UPDATE] Sora C++ SDK を 2023.4.0 に上げる
     - @melpon
 - [UPDATE] `CMake` を `3.25.1` に上げる
     - @voluntas