+++ CHANGES.md
@@ -35,6 +35,12 @@
     - @melpon
 - [ADD] `--max-retry` および `--retry-interval` オプションを追加
     - @melpon
+- [ADD] Lyra 向けオプションを追加
+    - `--sora-audio-codec-type` オプションに `LYRA` を追加
+    - `--sora-audio-codec-lyra-bitrate` オプションを追加
+    - `--sora-audio-codec-lyra-usedtx` オプションを追加
+    - `--sora-check-lyra-version` オプションを追加
+    - @torikizi
 - [FIX] 廃止になった `--sora-audio-opus-params-clock-rate` を削除する
     - @torikizi
 - [FIX] "data-channels" の "interval" 項目を指定するとエラーになる問題を修正